+++ crates/cast/src/main.rs
@@ -2,7 +2,8 @@ use crate::starknet_commands::account::Account;
 use crate::starknet_commands::{
     account, call::Call, declare::Declare, deploy::Deploy, invoke::Invoke, multicall::Multicall,
 };
-use anyhow::Result;
+use anyhow::{anyhow, Result};
+
 use camino::Utf8PathBuf;
 use cast::helpers::constants::{DEFAULT_ACCOUNTS_FILE, DEFAULT_MULTICALL_CONTENTS};
 use cast::helpers::scarb_utils::{parse_scarb_config, CastConfig};
@@ -206,8 +207,11 @@ async fn main() -> Result<()> {
         }
         Commands::Account(account) => match account.command {
             account::Commands::Add(add) => {
+                config.account = add.name.clone();
                 let mut result = starknet_commands::account::add::add(
-                    &config,
+                    &config.rpc_url,
+                    &config.account,
+                    &config.accounts_file,
                     &cli.path_to_scarb_toml,
                     &provider,
                     &add,
@@ -219,19 +223,22 @@ async fn main() -> Result<()> {
             }
             account::Commands::Create(create) => {
                 let chain_id = get_chain_id(&provider).await?;
-                if cli.keystore.is_none() {
-                    config.account = create.name;
+                if config.keystore == Utf8PathBuf::default() {
+                    config.account = create
+                        .name
+                        .ok_or_else(|| anyhow!("required argument --name not provided"))?;
                 }
                 let mut result = starknet_commands::account::create::create(
-                    &config,
+                    &config.rpc_url,
+                    &config.account,
+                    &config.accounts_file,
+                    &config.keystore,
                     &provider,
                     cli.path_to_scarb_toml,
                     chain_id,
                     create.salt,
                     create.add_profile,
                     create.class_hash,
-                    cli.keystore,
-                    cli.account.map(Utf8PathBuf::from),
                 )
                 .await;
 
@@ -240,13 +247,19 @@ async fn main() -> Result<()> {
             }
             account::Commands::Deploy(deploy) => {
                 let chain_id = get_chain_id(&provider).await?;
-                let keystore_path = Some(config.keystore).filter(|p| *p != Utf8PathBuf::default());
-                let account_path =
-                    Some(Utf8PathBuf::from(config.account)).filter(|p| *p != String::default());
+                let keystore_path =
+                    Some(config.keystore.clone()).filter(|p| *p != Utf8PathBuf::default());
+                let account_path = Some(Utf8PathBuf::from(config.account.clone()))
+                    .filter(|p| *p != String::default());
+                if config.keystore == Utf8PathBuf::default() {
+                    config.account = deploy
+                        .name
+                        .ok_or_else(|| anyhow!("required argument --name not provided"))?;
+                }
                 let mut result = starknet_commands::account::deploy::deploy(
                     &provider,
                     config.accounts_file,
-                    deploy.name,
+                    config.account,
                     chain_id,
                     deploy.max_fee,
                     cli.wait,