+++ crates/cast/src/starknet_commands/account/create.rs
@@ -3,10 +3,10 @@ use crate::starknet_commands::account::{
 };
 use anyhow::{anyhow, bail, Result};
 use camino::Utf8PathBuf;
-use cast::helpers::constants::OZ_CLASS_HASH;
+use cast::helpers::constants::{CREATE_KEYSTORE_PASSWORD_ENV_VAR, OZ_CLASS_HASH};
 use cast::helpers::response_structs::AccountCreateResponse;
 use cast::helpers::scarb_utils::CastConfig;
-use cast::{extract_or_generate_salt, get_chain_id, get_test_keystore_password, parse_number};
+use cast::{extract_or_generate_salt, get_chain_id, get_keystore_password, parse_number};
 use clap::Args;
 use serde_json::json;
 use starknet::accounts::{AccountFactory, OpenZeppelinAccountFactory};
@@ -20,8 +20,8 @@ use starknet::signers::{LocalWallet, SigningKey};
 #[command(about = "Create an account with all important secrets")]
 pub struct Create {
     /// Account name under which account information is going to be saved
-    #[clap(short, long, default_value = "")]
-    pub name: String,
+    #[clap(short, long)]
+    pub name: Option<String>,
 
     /// Salt for the address
     #[clap(short, long)]
@@ -38,15 +38,16 @@ pub struct Create {
 
 #[allow(clippy::too_many_arguments)]
 pub async fn create(
-    config: &CastConfig,
+    rpc_url: &str,
+    account: &str,
+    accounts_file: &Utf8PathBuf,
+    keystore: &Utf8PathBuf,
     provider: &JsonRpcClient<HttpTransport>,
     path_to_scarb_toml: Option<Utf8PathBuf>,
     chain_id: FieldElement,
     salt: Option<FieldElement>,
     add_profile: bool,
     class_hash: Option<String>,
-    keystore_path: Option<Utf8PathBuf>,
-    account_path: Option<Utf8PathBuf>,
 ) -> Result<AccountCreateResponse> {
     let salt = extract_or_generate_salt(salt);
     let class_hash = {
@@ -64,37 +65,30 @@ pub async fn create(
             .ok_or_else(|| anyhow!("Invalid address"))?,
     )?;
 
-    if let Some(keystore_path_) = &keystore_path {
-        let account_path_ = account_path.ok_or_else(|| {
-            anyhow!("--account must be passed and be a path when using --keystore")
-        })?;
+    if keystore == &Utf8PathBuf::default() {
+        write_account_to_accounts_file(account, accounts_file, chain_id, account_json.clone())?;
+    } else {
+        let account_path = Utf8PathBuf::from(&account);
+        if account_path == Utf8PathBuf::default() {
+            bail!("--account must be passed and be a path when using --keystore");
+        }
 
         let private_key = parse_number(
             account_json["private_key"]
                 .as_str()
                 .ok_or_else(|| anyhow!("Invalid private_key"))?,
         )?;
-        create_to_keystore(
-            private_key,
-            salt,
-            class_hash,
-            keystore_path_,
-            &account_path_,
-            add_profile,
-            config,
-            &path_to_scarb_toml,
-        )?;
-    } else {
-        write_account_to_accounts_file(
-            &path_to_scarb_toml,
-            &config.rpc_url,
-            &config.account,
-            &config.accounts_file,
-            &config.keystore,
-            chain_id,
-            account_json.clone(),
-            add_profile,
-        )?;
+        create_to_keystore(private_key, salt, class_hash, keystore, &account_path)?;
+    }
+
+    if add_profile {
+        let config = CastConfig {
+            rpc_url: rpc_url.into(),
+            account: account.into(),
+            accounts_file: accounts_file.into(),
+            keystore: keystore.into(),
+        };
+        add_created_profile_to_configuration(&path_to_scarb_toml, &config)?;
     }
 
     let mut output = vec![("address", format!("{address:#x}"))];
@@ -177,11 +171,8 @@ fn create_to_keystore(
     class_hash: FieldElement,
     keystore_path: &Utf8PathBuf,
     account_path: &Utf8PathBuf,
-    add_profile: bool,
-    config: &CastConfig,
-    path_to_scarb_toml: &Option<Utf8PathBuf>,
 ) -> Result<()> {
-    let password = get_test_keystore_password()?;
+    let password = get_keystore_password(CREATE_KEYSTORE_PASSWORD_ENV_VAR)?;
     let private_key = SigningKey::from_secret_scalar(private_key);
     private_key.save_as_keystore(keystore_path, &password)?;
 
@@ -199,9 +190,6 @@ fn create_to_keystore(
         }
     });
 
-    if add_profile {
-        add_created_profile_to_configuration(path_to_scarb_toml, config)?;
-    }
     write_account_to_file(&oz_account_json, account_path)
 }